Update test covarage
